+++ src/main/java/roomescape/domain/repository/JdbcReservationRepository.java
@@ -10,7 +10,6 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationRepository;
 
 @Repository
 public class JdbcReservationRepository implements ReservationRepository {
