+++ src/main/java/roomescape/domain/repository/ReservationRepository.java
@@ -1,6 +1,7 @@
-package roomescape.domain;
+package roomescape.domain.repository;
 
 import java.util.List;
+import roomescape.domain.Reservation;
 
 public interface ReservationRepository {
